quick fix, check that file exists before removing

+++ scn_fs/src/main/kotlin/GenerateGMTs.kt
@@ -32,7 +32,9 @@ fun generateGMTs(
         val outGmtFile = Paths.get(outDir, speciesName + GMT_PREFIX)
         val charset: Charset = StandardCharsets.UTF_8
 
-        Files.delete(outGmtFile)
+        if (Files.exists(outGmtFile)) {
+            Files.delete(outGmtFile)
+        }
         // Join files (lines)
         for (path in inputs) {
             val lines = Files.readAllLines(path, charset)
